+++ js/views/messageattachment.js
@@ -23,7 +23,9 @@ define(function(require) {
 	var $ = require('jquery');
 	var Handlebars = require('handlebars');
 	var Marionette = require('marionette');
+	var Radio = require('radio');
 	var MessageController = require('controller/messagecontroller');
+	var CalendarsPopoverView = require('views/calendarspopoverview');
 	var MessageAttachmentTemplate = require('text!templates/message-attachment.html');
 
 	/**
@@ -33,14 +35,32 @@ define(function(require) {
 		template: Handlebars.compile(MessageAttachmentTemplate),
 		ui: {
 			'downloadButton': '.attachment-download',
-			'saveToCloudButton': '.attachment-save-to-cloud'
+			'saveToCloudButton': '.attachment-save-to-cloud',
+			'importCalendarEventButton': '.attachment-import.calendar',
+			'attachmentImportPopover': '.attachment-import-popover'
 		},
 		events: {
-			'click': '_onDownload',
-			'click @ui.saveToCloudButton': '_onSaveToCloud'
+			'click': '_onClick',
+			'click @ui.saveToCloudButton': '_onSaveToCloud',
+			'click @ui.importCalendarEventButton': '_onImportCalendarEvent'
 		},
-		_onDownload: function(e) {
+		templateHelpers: function() {
+			return {
+				hasDavSupport: require('app').hasDavSupport
+			};
+		},
+		initialize: function() {
+			this.listenTo(Radio.ui, 'document:click', this._closeImportPopover);
+		},
+		_onClick: function(e) {
 			if (!e.isDefaultPrevented()) {
+				var $target = $(e.target);
+				if ($target.hasClass('select-calendar')) {
+					var url = $target.data('calendar-url');
+					this._uploadToCalendar(url);
+					return;
+				}
+
 				e.preventDefault();
 				window.location = this.model.get('downloadUrl');
 			}
@@ -57,16 +77,84 @@ define(function(require) {
 
 			// Loading feedback
 			this.ui.saveToCloudButton.removeClass('icon-folder')
-				.addClass('icon-loading-small')
-				.prop('disabled', true);
+					.addClass('icon-loading-small')
+					.prop('disabled', true);
 
 			var _this = this;
 			$.when(saving).always(function() {
 				// Remove loading feedback again
 				_this.ui.saveToCloudButton.addClass('icon-folder')
-					.removeClass('icon-loading-small')
-					.prop('disabled', false);
+						.removeClass('icon-loading-small')
+						.prop('disabled', false);
 			});
+		},
+		_onImportCalendarEvent: function(e) {
+			e.preventDefault();
+
+			this.ui.importCalendarEventButton
+					.removeClass('icon-add')
+					.addClass('icon-loading-small');
+
+			var fetchingCalendars = Radio.dav.request('calendars');
+
+			var _this = this;
+			$.when(fetchingCalendars).done(function(calendars) {
+				if (calendars.length > 0) {
+					_this.ui.attachmentImportPopover.removeClass('hidden');
+					var calendarsView = new CalendarsPopoverView({
+						collection: calendars
+					});
+					calendarsView.render();
+					_this.ui.attachmentImportPopover.html(calendarsView.$el);
+				} else {
+					Radio.ui.trigger('error:show', t('mail', 'No writable calendars found'));
+				}
+			});
+			$.when(fetchingCalendars).always(function() {
+				_this.ui.importCalendarEventButton
+						.removeClass('icon-loading-small')
+						.addClass('icon-add');
+			});
+		},
+		_uploadToCalendar: function(url) {
+			this._closeImportPopover();
+			this.ui.importCalendarEventButton
+					.removeClass('icon-add')
+					.addClass('icon-loading-small');
+
+			var downloadUrl = this.model.get('downloadUrl');
+			var downloadingAttachment = Radio.message.request('attachment:download', downloadUrl);
+
+			var _this = this;
+			$.when(downloadingAttachment).done(function(content) {
+
+				var importingCalendarEvent = Radio.dav.request('calendar:import', url, content);
+
+				$.when(importingCalendarEvent).fail(function() {
+					Radio.ui.trigger('error:show', t('mail', 'Error while importing the calendar event'));
+				});
+				$.when(importingCalendarEvent).always(function() {
+					_this.ui.importCalendarEventButton
+							.removeClass('icon-loading-small')
+							.addClass('icon-add');
+				});
+			});
+			$.when(downloadingAttachment.fail(function() {
+				Radio.ui.trigger('error:show', t('mail', 'Error while downloading calendar event'));
+				_this.ui.importCalendarEventButton
+						.removeClass('icon-loading-small')
+						.addClass('icon-add');
+			}));
+		},
+		_closeImportPopover: function(e) {
+			if (_.isUndefined(e)) {
+				this.ui.attachmentImportPopover.addClass('hidden');
+				return;
+			}
+			var $target = $(e.target);
+			if (this.$el.find($target).length === 0) {
+				this.ui.attachmentImportPopover.addClass('hidden');
+			}
 		}
 	});